Method to set app.engine to other templates with handlebars example
Detect an object as the 'view engine' option to change the render engine function.
A string will stay as the default behaviour of express.

The included `examples/template-handlebars` demonstrates how to
overwrite the template engine express uses (pug).

Include the express-handlebars module in the devDependencies so the
example will run correctly for users.